modularize homeroutes
